REFACTOR: Pull data classes out of main code files
